+++ src/main/java/com/dsd/rfoodsp/repository/UsuarioRepository.java
@@ -2,10 +2,9 @@ package com.dsd.rfoodsp.repository;
 
 import org.springframework.stereotype.Repository;
 
-import com.dsd.rfoodsp.entities.Usuario;
+import com.dsd.rfoodsp.model.entities.Usuario;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
 import java.util.Optional;
 
 @Repository
