+++ src/main/java/com/dsd/rfoodsp/service/RolService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.dsd.rfoodsp.entities.Rol;
+import com.dsd.rfoodsp.model.entities.Rol;
 import com.dsd.rfoodsp.repository.RolRepository;
 
 @Service
